mainブランチのセッションリストパフォーマンス改善を取り込み、
テストエラーを修正。

主な変更:
- first_user_message機能の取り込み (schema.sql, sessions.go)
- calculateFirstUserMessage()関数の追加
- パフォーマンス改善（757件で0.5秒未満）
- AvgTokens型の不一致を修正 (テスト)

+++ internal/db/project_stats_test.go
@@ -92,9 +92,9 @@ func TestGetProjectStats(t *testing.T) {
 		t.Errorf("Expected %d output tokens, got %d", expectedOutputTokens, stats.TotalOutputTokens)
 	}
 
-	expectedAvgTokens := (expectedInputTokens + expectedOutputTokens) / 3
+	expectedAvgTokens := float64(expectedInputTokens+expectedOutputTokens) / float64(3)
 	if stats.AvgTokens != expectedAvgTokens {
-		t.Errorf("Expected %d avg tokens, got %d", expectedAvgTokens, stats.AvgTokens)
+		t.Errorf("Expected %f avg tokens, got %f", expectedAvgTokens, stats.AvgTokens)
 	}
 
 	expectedErrorRate := float64(1) / float64(3) // 1セッションにエラーあり
